working kubernetes deploy for qc
