+++ stimela/backends/native/__init__.py
@@ -1,1 +1,1 @@
-from .run import run
+from .run_native import run
